+++ lib/src/first_open_slideshow_frame.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 import 'package:animations/animations.dart';
-import 'package:first_open_slideshow/slideshow_item_widget.dart';
+import 'package:first_open_slideshow/src/slideshow_page.dart';
+import 'package:first_open_slideshow/src/slideshow_page_widget.dart';
 import 'package:flare_splash_screen/flare_splash_screen.dart';
 import 'package:flutter/material.dart';
 
@@ -9,7 +10,7 @@ const double PAGE_ITEM_ICON_SIZE = 250;
 class FirstOpenSlideshowFrame extends StatefulWidget {
   static const String FIRST_RUN_PREF_KEY = "FIRST_RUN_PREF_KEY";
   static String stringForNext = "";
-  final List<PageItem> pageItems;
+  final List<SlideshowPage> pageItems;
   final VoidCallback finishCallback;
   final Duration animationDuration;
 
@@ -52,7 +53,7 @@ class _FirstOpenSlideshowFrameState extends State<FirstOpenSlideshowFrame> {
               },
               child: Container(
                 key: ValueKey<int>(this.currentPage),
-                child: PageItemWidget.fromPageItem(
+                child: SlideshowPageWidget.fromSlideshowPage(
                   widget.pageItems[currentPage],
                   () {
                     if (widget.pageItems[currentPage] ==
@@ -102,53 +103,6 @@ class _FirstOpenSlideshowFrameState extends State<FirstOpenSlideshowFrame> {
       () {
         showOverlay = true;
       },
-    );
-  }
-}
-
-class PageItem {
-  final String titleText;
-  final String captionText;
-  final String buttonText;
-  final Widget icon;
-  final Widget captionWidget;
-  final bool centerTitle;
-
-  ///Caption Text or Caption Widget are required
-  PageItem(
-      {@required this.titleText,
-      this.icon,
-      this.buttonText,
-      this.captionText,
-      this.centerTitle,
-      this.captionWidget});
-}
-
-class PageIconContainer extends StatelessWidget {
-  final Widget icon;
-
-  const PageIconContainer(this.icon, {Key key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: PAGE_ITEM_ICON_SIZE,
-      width: PAGE_ITEM_ICON_SIZE,
-      child: this.icon,
-    );
-  }
-}
-
-class PageText extends StatelessWidget {
-  final String text;
-
-  PageText(this.text);
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      this.text,
-      style: TextStyle(fontSize: 20),
     );
   }
 }