+++ CHANGELOG.md
@@ -9,3 +9,16 @@
 ## [0.0.3]
 
 * Added SharedAxisTransition from animations package
+
+## [0.0.4]
+
+* Minor Refactoring
+
+## [0.0.5]
+
+* **BREAKING**
+* Renamed PageItem to SlideshowPage
+* Renamed PageItemWidget to SlideshowPageWidget
+* Renamed PageText to SlideshowPageText
+* Renamed PageIconContainer to SlideshowPageIconContainer
+* Restructured directories